requirements.txt
pytest
playwright
pytest-playwright
python-dotenv
allure-pytest
logzero
pytest-xdist
faker
pyodbc
hypothesis
allure-python-commons